Fix for JRUBY-3357: Marshal load and dump rubyspec failures, part 2, Regexp

+++ src/org/jruby/RubyRegexp.java
@@ -2084,6 +2084,6 @@ public class RubyRegexp extends RubyObject implements ReOptions, EncodingCapable
     public static void marshalTo(RubyRegexp regexp, MarshalStream output) throws java.io.IOException {
         output.registerLinkTarget(regexp);
         output.writeString(new String(regexp.str.bytes,regexp.str.begin,regexp.str.realSize));
-        output.writeInt(regexp.pattern.getOptions() & EMBEDDABLE);
+        output.writeByte(regexp.pattern.getOptions() & EMBEDDABLE);
     }
 }
